pyproject.toml
[project]
name = "feature-extract-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "clickhouse-connect>=0.8.11",
    "dspy>=2.5.43",
    "google-generativeai>=0.8.3",
    "grpcio-tools>=1.68.1",
    "grpcio==1.69",
    "instructor[vertexai]>=1.7.0",
    "more-itertools>=10.5.0",
    "pymongo>=4.10.1",
    "sentence-transformers>=3.3.1",
    "toml>=0.10.2",
    "trafilatura>=2.0.0",
    "transformers",
    "vertexai>=1.71.1",
    "zstandard>=0.23.0",
    "aiohttp>=3.11.11",
    "polars>=1.19.0",
    "duckdb>=1.1.3",
    "python-dateutil>=2.9.0.post0",
]

[dependency-groups]
dev = [
    "feature-extractor-py",
    "pytest>=8.3.4",
]

[tool.uv.sources]
feature-extractor-py = { workspace = true }
transformers = { git = "https://github.com/huggingface/transformers.git" }
